Moved "ResourceType" and "DataUnitFactory" classes to be part of the DataUnitBase file, as they are more closely bound to them.
