+++ CART/CARTLib/core/DataUnitBase.py
@@ -1,9 +1,48 @@
 from abc import abstractmethod, ABC
 from pathlib import Path
-from typing import Any, Optional
+from typing import Any, Optional, Protocol, Mapping
 
 import slicer
+from slicer.i18n import tr as _
+
 from CARTLib.core.LayoutManagement import LayoutHandler
+
+
+class ResourceType(Protocol):
+    """
+    Protocol defining all behavior required for use within CART's cohort
+    editor GUI, as well as
+    """
+    # A unique identifier for this type internally
+    id: str
+    # How to display this type to the user (within GUIS)
+    pretty_name: str
+    # A human-readable description for this resource type
+    description: str = _(
+        "No description for this resource type was provided!"
+    )
+
+    @abstractmethod
+    def format_for_csv(self, resource_name: str) -> str:
+        """
+        Format a resource name string for use within a cohort CSV file.
+        """
+        ...
+
+    @abstractmethod
+    def format_for_gui(self, resource_name: str) -> str:
+        """
+        Format a resource name string for use within GUI elements.
+        """
+        ...
+
+    @abstractmethod
+    def is_type(self, csv_label: str) -> bool:
+        """
+        Return true if the passed label (taken from a cohort CSV file)
+        matches the formatting of this resource type (i.e. it could be
+        generated by the `format_for_csv` function defined above)
+        """
 
 
 class DataUnitBase(ABC):
@@ -169,3 +208,26 @@ class DataUnitBase(ABC):
     def __delete__(self, instance):
         # When the object is explicitly deleted, run cleaning first
         self.clean()
+
+
+class DataUnitFactory(Protocol):
+    """
+    Protocol signature which matches the elements of the DataUnitBase class
+    required for it to work well within CART's cohort system.
+    """
+    def __call__(
+        self,
+        case_data: dict[str, str],
+        data_path: Path,
+        scene: Optional[slicer.vtkMRMLScene] = None,
+    ) -> DataUnitBase:
+        # The "constructor"
+        ...
+
+    @abstractmethod
+    def resource_types(self) -> Mapping[str, ResourceType]:
+        """
+        Provide a map of identifiers to the resource type they represent
+        for this data unit type/factory.
+        """
+        ...
