+++ CART/CARTLib/core/TaskBaseClass.py
@@ -1,77 +1,15 @@
 import logging
 from abc import ABC, abstractmethod
-from pathlib import Path
-from typing import Generic, Optional, TypeVar, Protocol, Iterable, Mapping
+from typing import Generic, Optional, TypeVar
 
 import qt
-import slicer
 from slicer.i18n import tr as _
 
-from CARTLib.core.DataUnitBase import DataUnitBase
+from CARTLib.core.DataUnitBase import DataUnitBase, DataUnitFactory
 from CARTLib.utils.config import JobProfileConfig, MasterProfileConfig, DictBackedConfig
 
 # Generic type hint class for anything which is a subclass of DataUnitBase
 D = TypeVar("D", bound=DataUnitBase)
-
-
-class ResourceType(Protocol):
-    """
-    Protocol defining all behavior required for use within CART's cohort
-    editor GUI, as well as
-    """
-    # A unique identifier for this type internally
-    id: str
-    # How to display this type to the user (within GUIS)
-    pretty_name: str
-    # A human-readable description for this resource type
-    description: str = _(
-        "No description for this resource type was provided!"
-    )
-
-    @abstractmethod
-    def format_for_csv(self, resource_name: str) -> str:
-        """
-        Format a resource name string for use within a cohort CSV file.
-        """
-        ...
-
-    @abstractmethod
-    def format_for_gui(self, resource_name: str) -> str:
-        """
-        Format a resource name string for use within GUI elements.
-        """
-        ...
-
-    @abstractmethod
-    def is_type(self, csv_label: str) -> bool:
-        """
-        Return true if the passed label (taken from a cohort CSV file)
-        matches the formatting of this resource type (i.e. it could be
-        generated by the `format_for_csv` function defined above)
-        """
-
-
-class DataUnitFactory(Protocol):
-    """
-    Protocol signature which matches the elements of the DataUnitBase class
-    required for it to work well within CART's cohort system.
-    """
-    def __call__(
-        self,
-        case_data: dict[str, str],
-        data_path: Path,
-        scene: Optional[slicer.vtkMRMLScene] = None,
-    ) -> D:
-        # The "constructor"
-        ...
-
-    @abstractmethod
-    def resource_types(self) -> Mapping[str, ResourceType]:
-        """
-        Provide a map of identifiers to the resource type they represent
-        for this data unit type/factory.
-        """
-        ...
 
 
 class TaskBaseClass(ABC, Generic[D]):
@@ -172,34 +110,6 @@ class TaskBaseClass(ABC, Generic[D]):
         raise NotImplementedError("setup must be implemented in subclasses")
 
     ## Class Methods ##
-    @classmethod
-    def feature_types(cls, data_factory_label: str) -> dict[str, str]:
-        """
-        Provides a map containing a feature type label and its description
-        for a "class" of feature this task can handle for the given data factory.
-        `data_factory_label` will always be a value defined in `getDataUnitFactories`
-        above.
-
-        For example, if the given data unit factory can handle anatomical
-        volume files and segmentation labels, you might return the following:
-
-        ```
-        {
-            "Volume": "An anatomical volume you want to view. Must have 'volume' in its name.",
-            "Segmentation: "A segmentation label to overlay on viewed volumes. Must have 'segmentation' in its name"
-        }
-        ```
-
-        These are shown to the user during cohort creation, and used in conjunction
-        with `feature_label_for` below to ensure each cohort feature will abide by
-        the format required for this task when using the selected Data Unit Factory.
-
-        By default, this returns an empty dictionary; if CART sees this, it will not
-        provide ANY support to the user during cohort generation, likely result in
-        cohort files which are malformed and/or misleading.
-        """
-        return {}
-
     @classmethod
     def format_feature_label_for_type(
         cls, initial_label: str, data_unit_factory_type: str, feature_type: str
